// prismicio-types.d.ts
type Simplify<T> = { [KeyType in keyof T]: T[KeyType] };
// ...
export interface AssessmentsPageDocumentDataCreateAssessmentModalItem {
  /**
   * Title field in *Assessments page → Create assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.create_assessment_modal[].title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title: prismic.KeyTextField;

  /**
   * Description field in *Assessments page → Create assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.create_assessment_modal[].description
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  description: prismic.KeyTextField;

  /**
   * Title field label field in *Assessments page → Create assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Title
   * - **API ID Path**: assessments_page.create_assessment_modal[].title_field_label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title_field_label: prismic.KeyTextField;

  /**
   * Description field label field in *Assessments page → Create assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Description
   * - **API ID Path**: assessments_page.create_assessment_modal[].description_field_label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  description_field_label: prismic.KeyTextField;

  /**
   * Description field placeholder field in *Assessments page → Create assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.create_assessment_modal[].description_field_placeholder
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  description_field_placeholder: prismic.KeyTextField;

  /**
   * Submit button label field in *Assessments page → Create assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.create_assessment_modal[].submit_button_label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  submit_button_label: prismic.KeyTextField;
}

/**
 * Item in *Assessments page → Update assessment modal*
 */
export interface AssessmentsPageDocumentDataUpdateAssessmentModalItem {
  /**
   * Title field in *Assessments page → Update assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.update_assessment_modal[].title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  title: prismic.KeyTextField;

  /**
   * Description field in *Assessments page → Update assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.update_assessment_modal[].description
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  description: prismic.KeyTextField;

  /**
   * Submit button label field in *Assessments page → Update assessment modal*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.update_assessment_modal[].submit_button_label
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  submit_button_label: prismic.KeyTextField;
}

/**
 * Content for Assessments page documents
 */
interface AssessmentsPageDocumentData {
  /**
   * Search placeholder text field in *Assessments page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.search_placeholder
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  search_placeholder: prismic.KeyTextField;

  /**
   * No assessments message field in *Assessments page*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.no_assessments_message
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  no_assessments_message: prismic.KeyTextField;

  /**
   * Create assessment modal field in *Assessments page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.create_assessment_modal[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  create_assessment_modal: prismic.GroupField<
    Simplify<AssessmentsPageDocumentDataCreateAssessmentModalItem>
  >;

  /**
   * Update assessment modal field in *Assessments page*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: assessments_page.update_assessment_modal[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  update_assessment_modal: prismic.GroupField<
    Simplify<AssessmentsPageDocumentDataUpdateAssessmentModalItem>
  >;
}

/**
 * Assessments page document from Prismic
 *
 * - **API ID**: `assessments_page`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type AssessmentsPageDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithUID<
    Simplify<AssessmentsPageDocumentData>,
    "assessments_page",
    Lang
  >;

/**
 * Item in *Common translations → Button labels*
 */
export interface CommonTranslationsDocumentDataButtonLabelsItem {
  /**
   * Back to assessment board field in *Common translations → Button labels*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Back to assessment board
   * - **API ID Path**: common_translations.button_labels[].back_to_assessment_board
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  back_to_assessment_board: prismic.KeyTextField;

  /**
   * Yes, delete field in *Common translations → Button labels*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Yes, delete
   * - **API ID Path**: common_translations.button_labels[].yes__delete
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  yes__delete: prismic.KeyTextField;

  /**
   * Cancel field in *Common translations → Button labels*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Cancel
   * - **API ID Path**: common_translations.button_labels[].cancel
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  cancel: prismic.KeyTextField;
}

/**
 * Item in *Common translations → Error messages*
 */
export interface CommonTranslationsDocumentDataErrorMessagesItem {
  /**
   * General field in *Common translations → Error messages*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: common_translations.error_messages[].general
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  general: prismic.KeyTextField;

  /**
   * Impact validation field in *Common translations → Error messages*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: common_translations.error_messages[].impact_validation
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  impact_validation: prismic.KeyTextField;

  /**
   * Save assessment field in *Common translations → Error messages*
   *
   * - **Field Type**: Text
   * - **Placeholder**: *None*
   * - **API ID Path**: common_translations.error_messages[].save_assessment
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  save_assessment: prismic.KeyTextField;

  /**
   * Choose a unique title field in *Common translations → Error messages*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Please choose a unique title for your assessment
   * - **API ID Path**: common_translations.error_messages[].choose_a_unique_title
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  choose_a_unique_title: prismic.KeyTextField;
}

/**
 * Item in *Common translations → Success messages*
 */
export interface CommonTranslationsDocumentDataSuccessMessagesItem {
  /**
   * Save assessment field in *Common translations → Success messages*
   *
   * - **Field Type**: Text
   * - **Placeholder**: Assessment has been updated.
   * - **API ID Path**: common_translations.success_messages[].save_assessment
   * - **Documentation**: https://prismic.io/docs/field#key-text
   */
  save_assessment: prismic.KeyTextField;
}

/**
 * Content for Common translations documents
 */
interface CommonTranslationsDocumentData {
  /**
   * Button labels field in *Common translations*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: common_translations.button_labels[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  button_labels: prismic.GroupField<
    Simplify<CommonTranslationsDocumentDataButtonLabelsItem>
  >;

  /**
   * Error messages field in *Common translations*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: common_translations.error_messages[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  error_messages: prismic.GroupField<
    Simplify<CommonTranslationsDocumentDataErrorMessagesItem>
  >;

  /**
   * Success messages field in *Common translations*
   *
   * - **Field Type**: Group
   * - **Placeholder**: *None*
   * - **API ID Path**: common_translations.success_messages[]
   * - **Tab**: Main
   * - **Documentation**: https://prismic.io/docs/field#group
   */
  success_messages: prismic.GroupField<
    Simplify<CommonTranslationsDocumentDataSuccessMessagesItem>
  >;
}

/**
 * Common translations document from Prismic
 *
 * - **API ID**: `common_translations`
 * - **Repeatable**: `false`
 * - **Documentation**: https://prismic.io/docs/custom-types
 *
 * @typeParam Lang - Language API ID of the document.
 */
export type CommonTranslationsDocument<Lang extends string = string> =
  prismic.PrismicDocumentWithoutUID<
    Simplify<CommonTranslationsDocumentData>,
    "common_translations",
    Lang
  >;